* commit '8575bfb03a8312913b5573ab5fcf06fae218df4d [formerly 9e327afcb5aa8645ac8902fbb41e1b0306dc8bcb]': (121 commits)
  ODY-1455: Use ConversationsOptions in data handler
  ODY-1455: allow to define decimal separator, etc.
  ODY-1455: Replaced charset handling with broader ConversionOptions
  ODY-1455: New class for holding conversion options
  ODY-1455: Added getCharacter
  ODY-1454: Some fixes, create output on heartbeats
  ODY-1453: BSON Writing (Experimental)
  ODY-1453: XML2 ProtocolHandler working
  ODY-1452: fixed list and array handling
  ODY-1451: Handle both cases (String and number)
  ODY-1449: set done in case of ioexception
  Updated Feature descriptions
  Should be called XML Feature ...
  Server products need -data @user.home for cql2
  Lazy Loading in server, too
  Lazy Loading of cql2 parser
  ODY-1448: Updated Target Platform
  cleaned up logging, do not log per default to file
  ODY-1440: Remove ToTupleAO when input is already Tuple
  ODY-1440: Simplified isExecutable
  ...


Former-commit-id: 9f60fbbaf9bee559b2c30b0861fe9997e44b446a

+++ common/mep/src/de/uniol/inf/is/odysseus/mep/functions/string/StringMultiplicationOperator.java
@@ -30,7 +30,7 @@ public class StringMultiplicationOperator extends AbstractBinaryOperator<String>
     @Override
     public String getValue() {
 		String a = getInputValue(0);
-		Integer b = getInputValue(1);
+		Long b = getInputValue(1);
 		if ((a == null) || (b == null)) {
 			return null;
 		}
